Add files via upload
For this assignment, you are to
1) Read in a directed graph from an input text file,
2) Represent this graph in adjacency list form, and
3) Use the topological sort algorithm to either flag the graph as containing a cycle, or obtain a list of its vertices in
topologically sorted order.
The input file. Here is an example text-file representation of a graph:
5
Dallas Houston Austin Amarillo Lubbock
Dallas Austin 100
Houston Lubbock 100
Austin Lubbock 200
Dallas Amarillo 125
Lubbock Dallas 10
Austin Amarillo 220
This lists a graph with five vertices named Dallas, Houston, Austin, Amarillo, and Lubbock, and 6 edges (e.g., there is an
edge leading from Dallas to Austin, of cost 100).
Reading in the graph. You are to write a public class ReadGraph that supports the following public methods:
• public static String[ ] readVertices(Scanner fileIn).
This method returns a sorted list of vertex names (and retains a copy for future use).
• public static AdjListNode[ ] readEdges